Add index page and first recipie page
The index page will be used as a way to navigate between different
recipies. It will include links to different recipie pages. The
first one of these will be a recipie for chicken paprikash, a
Hungarian dish that I used to eat pretty frequently when I was
younger, but have not in a while.