Center images and images by height
Change the box type of the images on the page to block. This
will allow us to center the images on the page. Additionally,
change the sizing of the images to the height is consistently
300px and the width adjusts based on the size of the image.
In the future consider adjusting the height to be relative to
viewport height so the page displays similarly on smaller
screens.